+++ src/main/java/com/study/realworld/user/domain/Email.java
@@ -3,7 +3,7 @@ package com.study.realworld.user.domain;
 import static com.google.common.base.Preconditions.checkArgument;
 import static java.util.regex.Pattern.matches;
 
-import com.study.realworld.global.error.ErrorCode;
+import com.study.realworld.global.exception.ErrorCode;
 import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
